backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
email-validator==2.2.0
redis==5.0.8
slowapi==0.1.9
python-slugify==8.0.4
Pillow==10.4.0
aiofiles==24.1.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
pytest-cov==5.0.0
